+++ CHANGES.rst
@@ -1,5 +1,6 @@
 v0.3
 ====
+- Added support for units (.units, .constants), new serialization format.
 - Moved repository to github.com/chemreac/chemreac
 - Enabled use of Sundial's iterative linear solvers.
 - Added Incomplete LU preconditioner
@@ -9,7 +10,7 @@ v0.3
 - Removed unused code found by pyflakes
 - Updated logo and webpage
 - Updated setup.py for better automated releases.
-- Drop support for k_err and D_err attributes
+- Drop support for k_err and D_err attributes in ReactionDiffusion
 - Add support for longtable in chemreac.util.table.rsys2pdf_table(...)
 - Clarify if input data to plotting routines should be transformed or not.
 - Enable support for file objects in load/dump in serialization